requirements.txt
python-telegram-bot>=21.0
yt-dlp>=2024.12.1
python-dotenv>=1.0.0